Send metric methods were changed to have 1 thread for all metrics
